+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Current
+-------
+
+- Added ``Api.marshal_with()`` and ``Api.marshal_list_with()`` decorators
+- Added ``Api.marshal()`` shortcut
+
+
 0.1.1
 -----